+++ GD32F4xx_Demo_Suites_V2.6.1/GD32427V_START_Demo_Suites/Projects/02_GPIO_Key_Polling_mode/common/utilities.h
@@ -25,6 +25,6 @@
 int myatoi(const char* s);
 int myatohex(const char* s);
 unsigned int myatohexuint(const char* s);
-long int myatol(const char* s);
+long long int myatol(const char* s);
 
 #endif /* end of __UTILITIES_H__ */
